LayoutTests/imported/w3c/web-platform-tests/dom/events/event-global-set-before-handleEvent-lookup.any.worker-expected.txt


FAIL self.event is set before 'handleEvent' lookup assert_equals: expected (object) object "[object Event]" but got (undefined) undefined
